+++ test/unit/npm-utils.test.js
@@ -1,4 +1,4 @@
-const {getVersionList, resolveVersionRange} = require('../src/npm-utils');
+const {getVersionList, resolveVersionRange} = require('../../src/npm-utils');
 
 describe('getVersionList', () => {
   it('get the version list of an existing package', async () => {
